First working version for x64
Working for assemblies / types. Still some problems remaining with fields.

+++ src/HackF5.UnitySpy/Detail/ManagedClassInstance.cs
@@ -12,11 +12,11 @@
     [PublicAPI]
     public class ManagedClassInstance : ManagedObjectInstance
     {
-        private readonly uint definitionAddress;
+        private readonly long definitionAddress;
 
-        private readonly uint vtable;
+        private readonly long vtable;
 
-        public ManagedClassInstance([NotNull] AssemblyImage image, uint address)
+        public ManagedClassInstance([NotNull] AssemblyImage image, long address)
             : base(image, address)
         {
             if (image == null)
